+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### ✨ Novas Funcionalidades
 - Adicione novas funcionalidades aqui
 - Adicionada funcionaldiade de duplicação de transações.
+- Adicionada possibilidade de pular execução de recorrência.
 
 ### 🔧 Melhorias
 - Ajustado filtro de período do mês de vencimento das faturas do cartão para exibir a fatura atual.